requirements.txt
python-telegram-bot==13.7
pymongo==3.11.4
